Remove metric_metadata from internal_events
